Move PlayerSAO to dedicated files
