+++ src/database/database-files.cpp
@@ -20,11 +20,11 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include <cassert>
 #include <json/json.h>
 #include "database-files.h"
-#include "content_sao.h"
 #include "remoteplayer.h"
 #include "settings.h"
 #include "porting.h"
 #include "filesys.h"
+#include "server/player_sao.h"
 #include "util/string.h"
 
 // !!! WARNING !!!
