+++ src/network/serverpackethandler.cpp
@@ -20,7 +20,6 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "chatmessage.h"
 #include "server.h"
 #include "log.h"
-#include "content_sao.h"
 #include "emerge.h"
 #include "mapblock.h"
 #include "modchannels.h"
@@ -34,6 +33,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "network/connection.h"
 #include "network/networkprotocol.h"
 #include "network/serveropcodes.h"
+#include "server/player_sao.h"
 #include "util/auth.h"
 #include "util/base64.h"
 #include "util/pointedthing.h"
