+++ src/server/unit_sao.cpp
@@ -22,10 +22,6 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "scripting_server.h"
 #include "serverenvironment.h"
 
-/*
-	UnitSAO
- */
-
 UnitSAO::UnitSAO(ServerEnvironment *env, v3f pos) : ServerActiveObject(env, pos)
 {
 	// Initialize something to armor groups
